split context assembler into submodule
Split src/context/assembler.rs (1649 lines) into:
- assembler/mod.rs (518 lines) — core types, ContextAssembler impl, tests
- assembler/memory_selection.rs (518 lines) — memory selection pipeline
- assembler/view.rs (432 lines) — context view assembly, retrieval helpers
- assembler/compaction.rs (162 lines) — compaction metadata extraction

No public API changes. All re-exports through src/context/mod.rs unchanged.
